[SPT-777] New structure of functional tests
