+++ .github/workflows/test.yml
@@ -64,4 +64,3 @@ jobs:
           export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python
           # sh -ex ./snet/cli/test/utils/run_all_functional.sh
           python3 ./snet/cli/test/functional_tests/test_entry_point.py
-          python3 ./snet/cli/test/functional_tests/func_tests.py
